Split App into separate modules
