+++ app/javascript/src/admin/micro-clusters/App.jsx
@@ -1,19 +1,13 @@
 import React, { useEffect, useReducer, useContext } from "react";
 import Select from "react-select";
 import _ from "lodash";
-import Jsona from "jsona";
 
-import { getRequest } from "../../fetch";
 import { Spinner } from "../components/Spinner";
 import { DisplayMicroClusters } from "./DisplayMicroClusters";
 import { reducer, initalState } from "../components/clustering/reducer";
-import {
-  SET_LOADING_PERCENTAGE,
-  SET_MACRO_CLUSTERS,
-  SET_MICRO_CLUSTERS,
-  UPDATE_SELECTED_BRANDS
-} from "../components/clustering/actions";
+import { UPDATE_SELECTED_BRANDS } from "../components/clustering/actions";
 import { setInBrandSelector } from "../components/clustering/keyDownListener";
+import { loadMacroClusters, loadMicroClusters } from "./loadClusters";
 
 export const StateContext = React.createContext();
 export const DispatchContext = React.createContext();
@@ -114,78 +108,3 @@ const BrandSelector = () => {
     </div>
   );
 };
-const loadMicroClusters = (dispatch) => {
-  const formatter = new Jsona();
-  let data = [];
-  function run(page = 1) {
-    loadMicroClusterPage(page).then((json) => {
-      const next_page = json.meta.pagination.next_page;
-      // Remove clusters without collected inks
-      // Group collected inks
-      const pageData = formatter
-        .deserialize(json)
-        .filter((c) => c.collected_inks.length > 0)
-        .map((c) => {
-          const grouped_collected_inks = groupedInks(c.collected_inks);
-          return { ...c, grouped_collected_inks };
-        });
-      data = [...data, ...pageData];
-      if (next_page) {
-        run(next_page);
-      } else {
-        dispatch({ type: SET_MICRO_CLUSTERS, payload: data });
-      }
-    });
-  }
-  run();
-};
-
-const loadMicroClusterPage = (page) => {
-  return getRequest(
-    `/admins/micro_clusters.json?unassigned=true&without_ignored=true&page=${page}`
-  ).then((response) => response.json());
-};
-
-const loadMacroClusters = (dispatch) => {
-  let data = [];
-  const formatter = new Jsona();
-  function run(page = 1) {
-    loadMacroClusterPage(page).then((json) => {
-      const pagination = json.meta.pagination;
-      dispatch({
-        type: SET_LOADING_PERCENTAGE,
-        payload: (pagination.current_page * 100) / pagination.total_pages
-      });
-      const next_page = json.meta.pagination.next_page;
-      const pageData = formatter.deserialize(json).map((c) => {
-        const grouped_collected_inks = groupedInks(
-          c.micro_clusters.map((c) => c.collected_inks).flat()
-        );
-        return { ...c, grouped_collected_inks };
-      });
-      data = [...data, ...pageData];
-      if (next_page) {
-        run(next_page);
-      } else {
-        dispatch({ type: SET_MACRO_CLUSTERS, payload: data });
-      }
-    });
-  }
-  run();
-};
-
-const loadMacroClusterPage = (page) => {
-  return getRequest(`/admins/macro_clusters.json?page=${page}`).then(
-    (response) => response.json()
-  );
-};
-
-export const groupedInks = (collectedInks) =>
-  _.values(
-    _.mapValues(
-      _.groupBy(collectedInks, (ci) =>
-        ["brand_name", "line_name", "ink_name"].map((n) => ci[n]).join(",")
-      ),
-      (cis) => cis[0]
-    )
-  );
